Moved rDecorateSeedBiome(), rPopulatePreDecorate() & rPopulatePostDecorate() to BDRTG.
