+++ src/main/java/rtg/world/biome/BiomeDecoratorRTG.java
@@ -10,19 +10,23 @@ import net.minecraft.init.Blocks;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraft.world.biome.Biome;
+import net.minecraft.world.chunk.IChunkGenerator;
 import net.minecraft.world.gen.ChunkProviderSettings;
-import net.minecraft.world.gen.feature.WorldGenMinable;
-import net.minecraft.world.gen.feature.WorldGenerator;
+import net.minecraft.world.gen.feature.*;
 
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.terraingen.DecorateBiomeEvent;
 import net.minecraftforge.event.terraingen.OreGenEvent;
+import net.minecraftforge.event.terraingen.PopulateChunkEvent;
 import net.minecraftforge.event.terraingen.TerrainGen;
 
+import rtg.config.rtg.ConfigRTG;
 import rtg.util.CellNoise;
 import rtg.util.OpenSimplexNoise;
+import rtg.util.RandomUtil;
 import rtg.world.biome.realistic.RealisticBiomeBase;
 import rtg.world.gen.feature.WorldGenClayRTG;
+import static rtg.config.rtg.ConfigRTG.*;
 
 public class BiomeDecoratorRTG
 {
@@ -231,6 +235,159 @@ public class BiomeDecoratorRTG
                 }
             }
             return true;
+        }
+    }
+
+    /**
+     * When manually decorating biomes, sometimes you want the biome to partially decorate itself.
+     * That's what this method does... it calls the biome's decorate() method.
+     * If the conditions for decoration aren't met, we still need to generate ores.
+     */
+    public void rDecorateSeedBiome(World world, Random rand, int worldX, int worldZ, OpenSimplexNoise simplex, CellNoise cell, float strength, float river) {
+
+        if (strength > 0.3f) {
+            this.biome.decorate(world, rand, new BlockPos(worldX, 0, worldZ));
+        }
+        else {
+            this.decorateOres(world, rand, worldX, worldZ, simplex, cell, strength, river, false);
+        }
+    }
+
+    public void rPopulatePreDecorate(IChunkGenerator ichunkgenerator, World worldObj, Random rand, int chunkX, int chunkZ, boolean villageBuilding) {
+
+        int worldX = chunkX * 16;
+        int worldZ = chunkZ * 16;
+        boolean gen = true;
+
+        gen = TerrainGen.populate(ichunkgenerator, worldObj, rand, chunkX, chunkZ, villageBuilding, PopulateChunkEvent.Populate.EventType.LAKE);
+
+        // Underground water lakes.
+        if (ConfigRTG.enableWaterUndergroundLakes) {
+
+            if (gen && (ConfigRTG.waterUndergroundLakeChance > 0) && (rbb.waterUndergroundLakeChance > 0)) {
+
+                int i2 = worldX + rand.nextInt(16);// + 8;
+                int l4 = RandomUtil.getRandomInt(rand, 1, 50);
+                int i8 = worldZ + rand.nextInt(16);// + 8;
+
+                if (rand.nextInt(waterUndergroundLakeChance) == 0 && (RandomUtil.getRandomInt(rand, 1, waterUndergroundLakeChance) == 1)) {
+
+                    (new WorldGenLakes(Blocks.WATER)).generate(worldObj, rand, new BlockPos(new BlockPos(i2, l4, i8)));
+                }
+            }
+        }
+
+        // Surface water lakes.
+        if (ConfigRTG.enableWaterSurfaceLakes && !villageBuilding) {
+
+            if (gen && (waterSurfaceLakeChance > 0)) {
+
+                int i2 = worldX + rand.nextInt(16);// + 8;
+                int i8 = worldZ + rand.nextInt(16);// + 8;
+                int l4 = worldObj.getHeight(new BlockPos(i2, 0, i8)).getY();
+
+                //Surface lakes.
+                if (rand.nextInt(waterSurfaceLakeChance) == 0 && (RandomUtil.getRandomInt(rand, 1, waterSurfaceLakeChance) == 1)) {
+
+                    if (l4 > 63) {
+
+                        (new WorldGenLakes(Blocks.WATER)).generate(worldObj, rand, new BlockPos(i2, l4, i8));
+                    }
+                }
+            }
+        }
+
+        gen = TerrainGen.populate(ichunkgenerator, worldObj, rand, chunkX, chunkZ, villageBuilding, PopulateChunkEvent.Populate.EventType.LAVA);
+
+        // Underground lava lakes.
+        if (ConfigRTG.enableLavaUndergroundLakes) {
+
+            if (gen && (lavaUndergroundLakeChance > 0)) {
+
+                int i2 = worldX + rand.nextInt(16);// + 8;
+                int l4 = RandomUtil.getRandomInt(rand, 1, 50);
+                int i8 = worldZ + rand.nextInt(16);// + 8;
+
+                if (rand.nextInt(lavaUndergroundLakeChance) == 0 && (RandomUtil.getRandomInt(rand, 1, lavaUndergroundLakeChance) == 1)) {
+
+                    (new WorldGenLakes(Blocks.LAVA)).generate(worldObj, rand, new BlockPos(i2, l4, i8));
+                }
+            }
+        }
+
+        // Surface lava lakes.
+        if (ConfigRTG.enableLavaSurfaceLakes && !villageBuilding) {
+
+            if (gen && (lavaSurfaceLakeChance > 0)) {
+
+                int i2 = worldX + rand.nextInt(16);// + 8;
+                int i8 = worldZ + rand.nextInt(16);// + 8;
+                int l4 = worldObj.getHeight(new BlockPos(i2, 0, i8)).getY();
+
+                //Surface lakes.
+                if (rand.nextInt(lavaSurfaceLakeChance) == 0 && (RandomUtil.getRandomInt(rand, 1, lavaSurfaceLakeChance) == 1)) {
+
+                    if (l4 > 63) {
+
+                        (new WorldGenLakes(Blocks.LAVA)).generate(worldObj, rand, new BlockPos(i2, l4, i8));
+                    }
+                }
+            }
+        }
+
+        if (ConfigRTG.generateDungeons) {
+
+            gen = TerrainGen.populate(ichunkgenerator, worldObj, rand, chunkX, chunkZ, villageBuilding, PopulateChunkEvent.Populate.EventType.DUNGEON);
+
+            if (gen) {
+
+                for(int k1 = 0; k1 < ConfigRTG.dungeonFrequency; k1++) {
+
+                    int j5 = worldX + rand.nextInt(16);// + 8;
+                    int k8 = rand.nextInt(128);
+                    int j11 = worldZ + rand.nextInt(16);// + 8;
+
+                    (new WorldGenDungeons()).generate(worldObj, rand, new BlockPos(j5, k8, j11));
+                }
+            }
+        }
+    }
+
+    public void rPopulatePostDecorate(World worldObj, Random rand, int chunkX, int chunkZ, boolean flag) {
+
+        // Are flowing liquid modifications enabled?
+        if (!ConfigRTG.enableFlowingLiquidModifications) {
+            return;
+        }
+
+        int worldX = chunkX * 16;
+        int worldZ = chunkZ * 16;
+        int worldHeight = worldObj.provider.getActualHeight();
+
+        //Flowing water.
+        if (ConfigRTG.flowingWaterChance > 0) {
+            if (rand.nextInt(ConfigRTG.flowingWaterChance) == 0) {
+                for(int l18 = 0; l18 < 50; l18++)
+                {
+                    int l21 = worldX + rand.nextInt(16);// + 8;
+                    int k23 = rand.nextInt(rand.nextInt(worldHeight - 16) + 10);
+                    int l24 = worldZ + rand.nextInt(16);// + 8;
+                    (new WorldGenLiquids(Blocks.FLOWING_WATER)).generate(worldObj, rand, new BlockPos(l21, k23, l24));
+                }
+            }
+        }
+
+        //Flowing lava.
+        if (ConfigRTG.flowingLavaChance > 0) {
+            if (rand.nextInt(ConfigRTG.flowingLavaChance) == 0) {
+                for(int i19 = 0; i19 < 20; i19++)
+                {
+                    int i22 = worldX + rand.nextInt(16);// + 8;
+                    int l23 = rand.nextInt(worldHeight / 2);
+                    int i25 = worldZ + rand.nextInt(16);// + 8;
+                    (new WorldGenLiquids(Blocks.FLOWING_LAVA)).generate(worldObj, rand, new BlockPos(i22, l23, i25));
+                }
+            }
         }
     }
 }
